// DSDeviceDetection/domain/enums/commonconstants.ts
export const enum CommonConstants {
	ACCEPTVERSION = 'accept-version',
	ASYNC = 'Async',
	CANTBEDELETED = 'can\'t be deleted',
	CLIENTAUTHORIZATION = 'client-authorization',
	CLIENTAUTHENTICATION = 'client-authentication',
	CONTENTLENGTH = 'Content-Length',
	CONTENTTYPE = 'Content-Type',
	ERROR = 'Error',
	FALSE = 'false',
	IN = 'in',
	MULTI = 'multi',
	NOTFOUND = 'not found',
	OK = 'Ok',
	PARENTHESES = '()',
	POINT = '.',
	RECIEVED = 'Recieved',
	TRUE = 'true',
	TWOPOINTS = ':'
}